alert/requirements.txt
opencv-python-headless
numpy
pyyaml
requests
